+++ lbh15/properties/common_interface.py
@@ -1,0 +1,29 @@
+"""
+Module with the definition of common property interfaces"""
+from .interface import PropertyInterface
+
+
+class OxygenDiffusivityInterface(PropertyInterface):
+    """
+    Liquid metal *Oxygen diffusivity* property abstract class.
+    """
+    @property
+    def name(self) -> str:
+        """
+        str : Name of the property
+        """
+        return "o_dif"
+
+    @property
+    def units(self) -> str:
+        """
+        str : Oxygen diffusivity unit
+        """
+        return "[m^2.s^-1]"
+
+    @property
+    def long_name(self) -> str:
+        """
+        str : Oxygen diffusivity long name
+        """
+        return "oxygen diffusivity"
